Api_Python/requirements.txt
flask==2.0.3
mysql-connector-python==8.0.28
werkzeug<2.3
requests>=2.0.0